+++ Assets/Adjust/Unity/AdjustUtils.cs
@@ -3,10 +3,8 @@ using System.Collections.Generic;
 
 using UnityEngine;
 
-namespace com.adjust.sdk
-{
-    public class AdjustUtils
-    {
+namespace com.adjust.sdk {
+    public class AdjustUtils {
         #region Constants
         public static string KeyAdid = "adid";
         public static string KeyMessage = "message";
@@ -24,81 +22,62 @@ namespace com.adjust.sdk
         #endregion
 
         #region Public methods
-        public static int ConvertLogLevel (AdjustLogLevel? logLevel)
-        {
-            if (logLevel == null)
-            {
+        public static int ConvertLogLevel(AdjustLogLevel? logLevel) {
+            if (logLevel == null) {
                 return -1;
             }
 
             return (int)logLevel;
         }
 
-        public static int ConvertBool (bool? value)
-        {
-            if (value == null)
-            {
+        public static int ConvertBool(bool? value) {
+            if (value == null) {
                 return -1;
             }
 
-            if (value.Value)
-            {
+            if (value.Value) {
                 return 1;
-            }
-            else
-            {
+            } else {
                 return 0;
             }
         }
 
-        public static double ConvertDouble (double? value)
-        {
-            if (value == null)
-            {
+        public static double ConvertDouble(double? value) {
+            if (value == null) {
                 return -1;
             }
 
             return (double)value;
         }
 
-        public static string ConvertListToJson (List<String> list)
-        {
-            if (list == null)
-            {
+        public static string ConvertListToJson(List<String> list) {
+            if (list == null) {
                 return null;
             }
 
-            var jsonArray = new JSONArray ();
+            var jsonArray = new JSONArray();
             
-            foreach (var listItem in list)
-            {
-                jsonArray.Add (new JSONData (listItem));
+            foreach (var listItem in list) {
+                jsonArray.Add(new JSONData(listItem));
             }
 
-            return jsonArray.ToString ();
+            return jsonArray.ToString();
         }
 
-        public static string GetJsonResponseCompact (Dictionary<string, object> dictionary)
-        {
+        public static string GetJsonResponseCompact(Dictionary<string, object> dictionary) {
             string logJsonResponse = "";
 
-            if (dictionary == null)
-            {
+            if (dictionary == null) {
                 return logJsonResponse;
-            }
-            else
-            {
+            } else {
                 int preLoopCounter = 0;
                 logJsonResponse += "{";
 
-                foreach (KeyValuePair<string, object> pair in dictionary)
-                {
+                foreach (KeyValuePair<string, object> pair in dictionary) {
                     String valueString = pair.Value as string;
 
-                    if (valueString != null)
-                    {
-                        if (++preLoopCounter > 1)
-                        {
+                    if (valueString != null) {
+                        if (++preLoopCounter > 1) {
                             logJsonResponse += ",";
                         }
 
@@ -109,13 +88,12 @@ namespace com.adjust.sdk
 
                     Dictionary<string, object> valueDictionary = pair.Value as Dictionary<string, object>;
 
-                    if (++preLoopCounter > 1)
-                    {
+                    if (++preLoopCounter > 1) {
                         logJsonResponse += ",";
                     }
 
                     logJsonResponse += "\"" + pair.Key + "\"" + ":";
-                    logJsonResponse += GetJsonResponseCompact (valueDictionary);
+                    logJsonResponse += GetJsonResponseCompact(valueDictionary);
                 }
 
                 logJsonResponse += "}";
@@ -124,49 +102,43 @@ namespace com.adjust.sdk
             return logJsonResponse;
         }
 
-        public static String GetJsonString (JSONNode node, string key)
-        {
-            if (node == null)
-            {
+        public static String GetJsonString(JSONNode node, string key) {
+            if (node == null) {
                 return null;
             }
 
             // Access value object and cast it to JSONData.
-            var nodeValue = node [key] as JSONData;
+            var nodeValue = node[key] as JSONData;
 
-            if (nodeValue == null)
-            {
+            if (nodeValue == null) {
                 return null;
             }
 
             return nodeValue.Value;
         }
 
-        public static void WriteJsonResponseDictionary (JSONClass jsonObject, Dictionary<string, object> output)
-        {
-            foreach (KeyValuePair<string, JSONNode> pair in jsonObject)
-            {
+        public static void WriteJsonResponseDictionary(JSONClass jsonObject, Dictionary<string, object> output) {
+            foreach (KeyValuePair<string, JSONNode> pair in jsonObject) {
                 // Try to cast value as a complex object.
                 var subNode = pair.Value.AsObject;
                 var key = pair.Key;
 
                 // Value is not a complex object.
-                if (subNode == null)
-                {
+                if (subNode == null) {
                     var value = pair.Value.Value;
-                    output.Add (key, value);
+                    output.Add(key, value);
 
                     continue;
                 }
 
                 // Create new dictionary for complex type.
-                var newSubDictionary = new Dictionary<string, object> ();
+                var newSubDictionary = new Dictionary<string, object>();
 
                 // Save it in the current dictionary.
-                output.Add (key, newSubDictionary);
+                output.Add(key, newSubDictionary);
                 
                 // Recursive call to fill new dictionary.
-                WriteJsonResponseDictionary (subNode, newSubDictionary);
+                WriteJsonResponseDictionary(subNode, newSubDictionary);
             }
         }
         #endregion